pyproject.toml
[tool.poetry]
name = "snooze-server"
version = "1.4.2"
description = "Monitoring tool for logs aggregation and alerting"
authors = [
    "Florian Dematraz <florian.dematraz@snoozeweb.net>",
    "Guillaume Ludinard <guillaume.ludi@gmail.com>",
]
license = "AGPL-3.0-or-later"
readme = "README.md"
packages = [
    { include = "snooze" },
]

[tool.poetry.scripts]
snooze-server = "snooze.__main__:main"
snooze = "snooze.cli.__main__:snooze"
check_snooze_server = "snooze.cli.health:check_snooze_server"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
falcon = "^3.0.1"
Jinja2 = "^3.0.1"
ldap3 = "^2.9.1"
MarkupSafe = "^2.0.1"
pathlib = "^1.0.1"
prometheus-client = "^0.13.1"
PyJWT = "^2.3.0"
pymongo = "3.12.1"
pyparsing = "^2.4.7"
python-dateutil = "^2.8.2"
PyYAML = "5.4.1"
requests-unixsocket = "^0.2.0"
requests = "^2.26.0"
tinydb = "4.5.2"
waitress = "^2.0.0"
netifaces = "^0.11.0"
kombu = "^5.1.0"
pydantic = "^1.9.0"

[tool.poetry.group.lint.dependencies]
pylint = "^2.13.5"
pyright = "^1.1.238"
pylint-pytest = "^1.1.2"
pylint-pydantic = "^0.1.4"

[tool.poetry.group.security.dependencies]
bandit = "^1.7.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"
pytest-cov = "^2.12.1"
pytest-data = "^0.4"
pytest-mongodb = "^2.2.0"
pytest-spec = "^3.2.0"
python-mimeparse = "^1.6.0"
pytest-sugar = "^0.9.4"
pytest-timeout = "^1.4.2"
freezegun = "^1.1.0"
responses = "^0.20.0"

[tool.poetry.group.build.dependencies]
virtualenv = "^20.8.1"
invoke = "^1.7.0"
toml = "^0.10.2"

[tool.pylint.MASTER]
load-plugins = ['pylint_pytest', 'pylint_pydantic']
extension-pkg-whitelist = ['falcon', 'netifaces', 'pydantic']

[tool.pylint."MESSAGE CONTROL"]
disable = ['no-else-return']

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pyright]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
testpaths = "tests"
log_level = "DEBUG"
log_cli_format = "%(asctime)s %(name)-20s %(levelname)-8s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core>=1.1.0a6"]
build-backend = "poetry.core.masonry.api"
